TRITON-1934: update eslint (#8)
This updates eslint to the minimum version necessary to remediate
the eslint security vulnerability pointed out by GitHub. It was
also necessary to update node-eslint-plugin-joyent which
which previously did not support eslint v4.x.x. This update came
with a few new rules that caused make check to fail. In order for
make check to pass without substantial changes there were a few
minor changes and exclusions.

to a version without a security vulnerability